pyproject.toml
[project]
name = "filexa"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.4",
    "fpdf2>=2.8.7",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pdf2docx>=0.5.8",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.1",
    "pillow-heif>=1.3.0",
    "playwright>=1.58.0",
    "pymupdf>=1.26.7",
    "pypdf>=6.6.2",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-multipart>=0.0.22",
    "python-pptx>=1.0.2",
    "reportlab>=4.4.10",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
